+++ Sources/EmojiKit/Views/EmojiGrid.swift
@@ -31,9 +31,10 @@ public struct EmojiGrid<ItemView: View, SectionView: View>: View {
     ///   - categories: The categories to list, by default `.standard`.
     ///   - query: The search query to apply, if any.
     ///   - selection: The current grid selection, if any.
-    ///   - persistedCategory: The persisted category to affect when picking any emoji, if any.
+    ///   - persistedCategory: A persisted category to append to when picking an emoji, if any.
     ///   - geometryProxy: An optional geometry proxy, required to perform arrow/move-based navigation.
-    ///   - action: An action to trigger when an emoji is tapped or picked.
+    ///   - action: An action to trigger when an emoji is tapped or picked, if any.
+    ///   - categoryEmojis: An optional function that can determine which emojis to show for a certain category, by default all.
     ///   - section: A grid section title view builder.
     ///   - item: A grid item view builder.
     public init(
@@ -45,6 +46,7 @@ public struct EmojiGrid<ItemView: View, SectionView: View>: View {
         persistedCategory: EmojiCategory.PersistedCategory? = nil,
         geometryProxy: GeometryProxy? = nil,
         action: @escaping (Emoji) -> Void = { _ in },
+        categoryEmojis: @escaping (EmojiCategory) -> [Emoji] = { $0.emojis },
         @ViewBuilder section: @escaping (Emoji.GridSectionParameters) -> SectionView,
         @ViewBuilder item: @escaping (Emoji.GridItemParameters) -> ItemView
     ) {
@@ -57,6 +59,7 @@ public struct EmojiGrid<ItemView: View, SectionView: View>: View {
         self.persistedCategory = persistedCategory
         self.geometryProxy = geometryProxy
         self.action = action
+        self.categoryEmojis = categoryEmojis
         self.section = section
         self.item = item
         self._selection = selection
@@ -68,6 +71,7 @@ public struct EmojiGrid<ItemView: View, SectionView: View>: View {
     private let persistedCategory: EmojiCategory.PersistedCategory?
     private let geometryProxy: GeometryProxy?
     private let action: (Emoji) -> Void
+    private let categoryEmojis: (EmojiCategory) -> [Emoji]
     private let section: (Emoji.GridSectionParameters) -> SectionView
     private let item: (Emoji.GridItemParameters) -> ItemView
 
@@ -326,13 +330,13 @@ private extension EmojiGrid {
     func emojis(
         for category: EmojiCategory
     ) -> [Emoji] {
-        category.emojis
+        categoryEmojis(category)
     }
     
     func hasEmojis(
         for category: EmojiCategory
     ) -> Bool {
-        !category.emojis.isEmpty
+        !emojis(for: category).isEmpty
     }
     
     func isSelected(
@@ -363,6 +367,7 @@ private extension EmojiGrid {
                         axis: axis,
                         query: query,
                         selection: $selection,
+                        categoryEmojis: { Array($0.emojis.prefix(50)) },
                         section: { $0.view },
                         item: { $0.view }
                     )