+++ RELEASE_NOTES.md
@@ -13,6 +13,8 @@ EmojiKit tries to honor semantic versioning:
 
 This version optimizes emoji search and adjusts some localized texts.
 
+This version also renames some functions to use `in locale` instead of `for locale` terminology.
+
 
 
 ## 1.1